Fix text for editor page navigation and title
