+++ frontend/src/pages/OperatorEditorPage/OperatorYamlEditorPage.js
@@ -1,22 +1,51 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
+import classNames from 'classnames';
 import { connect } from 'react-redux';
+import * as _ from 'lodash-es';
 import { helpers } from '../../common/helpers';
 import { reduxConstants } from '../../redux';
 import YamlViewer from '../../components/YamlViewer';
-import OperatorEditorPage from './OperatorEditorPage';
 import { normalizeYamlOperator, yamlFromOperator } from './editorPageUtils';
+import OperatorEditorSubPage from './OperatorEditorSubPage';
+import PreviewOperatorModal from '../../components/modals/PreviewOperatorModal';
+import { defaultOperator } from '../../utils/operatorUtils';
 
 class OperatorYamlEditorPage extends React.Component {
   state = {
     yaml: '',
-    yamlError: null
+    yamlError: null,
+    previewShown: false
   };
 
   componentDidMount() {
     const { operator } = this.props;
     this.updateYaml(operator);
   }
+
+  backToPackageYourOperator = e => {
+    e.preventDefault();
+    this.props.history.push('/editor');
+  };
+
+  hidePreviewOperator = () => {
+    this.setState({ previewShown: false });
+  };
+
+  showPreviewOperator = () => {
+    this.setState({ previewShown: true });
+  };
+
+  doClearContents = () => {
+    const { resetEditorOperator, hideConfirmModal } = this.props;
+    resetEditorOperator();
+    hideConfirmModal();
+  };
+
+  clearContents = () => {
+    const { showConfirmModal } = this.props;
+    showConfirmModal(this.doClearContents);
+  };
 
   updateYaml = operator => {
     let yaml;
@@ -50,16 +79,41 @@ class OperatorYamlEditorPage extends React.Component {
     </div>
   );
 
-  render() {
-    const { history } = this.props;
-    const { yaml, yamlError } = this.state;
+  renderButtonBar() {
+    const { operator } = this.props;
+    const { yamlError } = this.state;
+
+    const isDefault = _.isEqual(operator, defaultOperator);
+
+    const previewClasses = classNames('oh-button oh-button-secondary', { disabled: yamlError });
+    const clearClasses = classNames('oh-button oh-button-secondary', { disabled: isDefault });
 
     return (
-      <OperatorEditorPage
+      <div className="oh-operator-editor-page__button-bar">
+        <div>
+          <button className="oh-button oh-button-secondary" onClick={this.backToPackageYourOperator}>
+            Back to Package your Operator
+          </button>
+          <button className={previewClasses} disabled={yamlError} onClick={this.showPreviewOperator}>
+            Preview
+          </button>
+        </div>
+        <button className={clearClasses} disabled={isDefault} onClick={this.clearContents}>
+          Clear Content
+        </button>
+      </div>
+    );
+  }
+
+  render() {
+    const { history, operator } = this.props;
+    const { yaml, yamlError, previewShown } = this.state;
+
+    return (
+      <OperatorEditorSubPage
         title="YAML Editor"
         header={this.renderHeader()}
-        isForm={false}
-        okToPreview={!yamlError}
+        buttonBar={this.renderButtonBar()}
         history={history}
       >
         <YamlViewer
@@ -69,7 +123,8 @@ class OperatorYamlEditorPage extends React.Component {
           error={yamlError}
           allowClear={false}
         />
-      </OperatorEditorPage>
+        <PreviewOperatorModal show={previewShown} yamlOperator={operator} onClose={this.hidePreviewOperator} />
+      </OperatorEditorSubPage>
     );
   }
 }
@@ -77,6 +132,9 @@ class OperatorYamlEditorPage extends React.Component {
 OperatorYamlEditorPage.propTypes = {
   operator: PropTypes.object,
   storeEditorOperator: PropTypes.func,
+  resetEditorOperator: PropTypes.func,
+  showConfirmModal: PropTypes.func,
+  hideConfirmModal: PropTypes.func,
   history: PropTypes.shape({
     push: PropTypes.func.isRequired
   }).isRequired
@@ -84,7 +142,10 @@ OperatorYamlEditorPage.propTypes = {
 
 OperatorYamlEditorPage.defaultProps = {
   operator: {},
-  storeEditorOperator: helpers.noop
+  storeEditorOperator: helpers.noop,
+  resetEditorOperator: helpers.noop,
+  showConfirmModal: helpers.noop,
+  hideConfirmModal: helpers.noop
 };
 
 const mapDispatchToProps = dispatch => ({
@@ -92,6 +153,23 @@ const mapDispatchToProps = dispatch => ({
     dispatch({
       type: reduxConstants.SET_EDITOR_OPERATOR,
       operator
+    }),
+  resetEditorOperator: () =>
+    dispatch({
+      type: reduxConstants.RESET_EDITOR_OPERATOR
+    }),
+  showConfirmModal: onConfirm =>
+    dispatch({
+      type: reduxConstants.CONFIRMATION_MODAL_SHOW,
+      title: 'Clear Content',
+      heading: <span>Are you sure you want to clear the current content of the editor?</span>,
+      confirmButtonText: 'Clear',
+      cancelButtonText: 'Cancel',
+      onConfirm
+    }),
+  hideConfirmModal: () =>
+    dispatch({
+      type: reduxConstants.CONFIRMATION_MODAL_HIDE
     })
 });
